pyproject.toml
[project]
name = "hort"
version = "0.1.0"
description = "Remote window viewer - watch your machine from your phone"
requires-python = ">=3.12,<3.15"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.34,<1.0",
    "Pillow>=11.0,<12.0",
    "qrcode[pil]>=8.0,<9.0",
    "pydantic>=2.10,<3.0",
    "pyobjc-framework-Quartz>=11.0; sys_platform == 'darwin'",
    "pyobjc-framework-ApplicationServices>=12.1,<13.0; sys_platform == 'darwin'",
    "websockets (>=16.0,<17.0)",
    "itsdangerous (>=2.2.0,<3.0.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "psutil (>=7.2.2,<8.0.0)",
    "aiortc (>=1.14.0,<2.0.0)",
    "rich (>=14.3.3,<15.0.0)",
    "click (>=8.3.1,<9.0.0)",
]
# pymongo installed separately (not in lock file to avoid Docker build issues)
# pip install pymongo

[project.scripts]
hort = "hort.cli:main"
claude-chat = "llmings.llms.claude_code.__main__:main"

[tool.poetry]
packages = [{include = "hort"}, {include = "llmings"}]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0,<9.0"
pytest-cov = ">=6.0,<7.0"
pytest-asyncio = ">=0.24,<1.0"
httpx = ">=0.28,<1.0"
mypy = ">=1.14,<2.0"
llming-com = {path = "../llming-com", develop = true}
playwright = "^1.58.0"
pytest-playwright = "^0.7.2"
aiogram = "^3.26.0"
llming-models = {path = "../llming-models", markers = "python_version >= '3.14'"}
office-connect = {path = "../office-connect", develop = true, markers = "python_version >= '3.14'"}


[tool.poetry.group.docs.dependencies]
mkdocs-material = "^9.7.6"

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
exclude = ["tests/"]

[[tool.mypy.overrides]]
module = "llming_com.*"
ignore_missing_imports = true
follow_untyped_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = ["integration: tests that hit real macOS APIs (run with -m integration)"]
addopts = "-m 'not integration'"
timeout = 60

[tool.coverage.run]
source = ["hort"]
omit = ["hort/static/*", "llmings/*", "hort/terminal.py", "hort/termd.py", "hort/termd_client.py", "hort/access/*", "hort/config.py", "hort/plugins.py"]

[tool.coverage.report]
fail_under = 100
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "def _handle_token_",
    "def _handle_tunnel_status",
    "def verify_token",
    "token_create",
    "token_list",
    "token_verify",
    "token_revoke",
    "tunnel_status",
    "_internal/verify-token",
    "if TYPE_CHECKING",
    "logger\\.warning",
    "logger\\.info",
    "get_store",
    "api/config",
    "_start_cloud_connector",
    "hort.access.tunnel_client",
]

[build-system]
requires = ["poetry-core>=2.0"]
build-backend = "poetry.core.masonry.api"
